samples/requirements.txt
python-socketio
requests # required by socketio
websocket-client # requested by socketio client
Flask
